user example initialize fix + header reorganization
- fixed initialize in example
- reorganized headers
