+++ cmake/Modules/KokkosPythonOptions.cmake
@@ -3,16 +3,6 @@
 #
 INCLUDE(KokkosPythonUtilities)
 
-# don't build a static python module
-SET(BUILD_SHARED_LIBS ON CACHE BOOL "Build shared libraries")
-
-# force to release if not specified
-IF("${CMAKE_BUILD_TYPE}" STREQUAL "")
-    SET(CMAKE_BUILD_TYPE Release CACHE STRING "Build type" FORCE)
-ENDIF()
-
-SET(CMAKE_INSTALL_RPATH_USE_LINK_PATH ON CACHE BOOL "Install with rpath")
-
 # backwards compat
 IF(NOT DEFINED ENABLE_EXAMPLES)
     SET(BUILD_EXAMPLES OFF CACHE BOOL "(deprecated) Use ENABLE_EXAMPLES")
@@ -43,9 +33,10 @@ SET(_VIEW_RANK_MSG "Set this value to the max number of ranks needed for Kokkos:
 ADD_FEATURE(CMAKE_BUILD_TYPE "Build type")
 ADD_FEATURE(CMAKE_INSTALL_PREFIX "Installation prefix")
 ADD_FEATURE(CMAKE_CXX_FLAGS "C++ compiler flags")
-ADD_OPTION(CMAKE_INSTALL_RPATH_USE_LINK_PATH "Build with rpath resolution" ON)
 ADD_FEATURE(Kokkos_CXX_STANDARD "Kokkos C++ Standard")
 ADD_FEATURE(Kokkos_DIR "Kokkos installation")
+
+ADD_OPTION(CMAKE_INSTALL_RPATH_USE_LINK_PATH "Install with rpath to linked libraries" ON)
 ADD_OPTION(ENABLE_INTERNAL_PYBIND11 "Build with pybind11 submodule" ON)
 ADD_OPTION(ENABLE_INTERNAL_KOKKOS "Build with kokkos submodule" ${_INTERNAL_KOKKOS})
 ADD_OPTION(ENABLE_WERROR "Build with -Werror" OFF)
@@ -61,6 +52,12 @@ ADD_OPTION(ENABLE_LAYOUTS "Build support for layouts (long NVCC compile times)"
     ${_ENABLE_MEM_DEFAULT})
 ADD_OPTION(ENABLE_MEMORY_TRAITS "Build support for memory traits (long NVCC compile times)"
     ${_ENABLE_LAY_DEFAULT})
+ADD_OPTION(CMAKE_UNITY_BUILD "Enable unity build" ON)
+
+SET(CMAKE_UNITY_BUILD_BATCH_SIZE 6 CACHE STRING "Unity build batch size")
+IF(CMAKE_UNITY_BUILD)
+    ADD_FEATURE(CMAKE_UNITY_BUILD_BATCH_SIZE "Unity build batch size")
+ENDIF()
 
 SET(ENABLE_VIEW_RANKS "4" CACHE STRING "${_VIEW_RANK_MSG}")
 IF(ENABLE_VIEW_RANKS LESS 0 OR ENABLE_VIEW_RANKS GREATER 7)
@@ -121,18 +118,6 @@ SET(Python3_ARTIFACTS_INTERACTIVE ON CACHE BOOL "Create CMake cache entries so t
 
 IF(DEFINED PYTHON_EXECUTABLE AND NOT DEFINED Python3_EXECUTABLE)
     SET(Python3_EXECUTABLE ${PYTHON_EXECUTABLE})
-ENDIF()
-
-# always disallow unity build
-IF(ENABLE_MEMORY_TRAITS)
-    ADD_OPTION(CMAKE_UNITY_BUILD "Enable unity build" ON)
-ELSE()
-    ADD_FEATURE(CMAKE_UNITY_BUILD "Enable unity build")
-ENDIF()
-
-SET(CMAKE_UNITY_BUILD_BATCH_SIZE 6 CACHE STRING "Unity build batch size")
-IF(CMAKE_UNITY_BUILD)
-    ADD_FEATURE(CMAKE_UNITY_BUILD_BATCH_SIZE "Unity build batch size")
 ENDIF()
 
 OPTION(ENABLE_CTP "Enable compile-time-perf" OFF)
