+++ cmake/Modules/KokkosPythonUtilities.cmake
@@ -205,6 +205,12 @@ ENDFUNCTION()
 #          Add an option and add as a feature if NO_FEATURE is not provided
 #
 FUNCTION(ADD_OPTION _NAME _MESSAGE _DEFAULT)
+    GET_PROPERTY(_ADDED GLOBAL PROPERTY ${PROJECT_NAME}_ADDED_OPTIONS)
+    IF("${_NAME}" IN_LIST _ADDED)
+        RETURN()
+    ENDIF()
+
+    SET_PROPERTY(GLOBAL APPEND PROPERTY ${PROJECT_NAME}_ADDED_OPTIONS ${_NAME})
     OPTION(${_NAME} "${_MESSAGE}" ${_DEFAULT})
     IF("NO_FEATURE" IN_LIST ARGN)
         MARK_AS_ADVANCED(${_NAME})
