+++ include/traits.hpp
@@ -44,12 +44,14 @@
 
 #pragma once
 
-#include "Kokkos_Core_fwd.hpp"
-#include "Kokkos_Layout.hpp"
-#include "Kokkos_MemoryTraits.hpp"
-
 #include "common.hpp"
+#include "concepts.hpp"
+#include "defines.hpp"
 #include "fwd.hpp"
+
+#include <Kokkos_Core_fwd.hpp>
+#include <Kokkos_Layout.hpp>
+#include <Kokkos_MemoryTraits.hpp>
 
 //--------------------------------------------------------------------------------------//
