+++ src/backend_version.cpp
@@ -42,8 +42,12 @@
 //@HEADER
 */
 
-#include "Kokkos_Core_fwd.hpp"
-#include "libpykokkos.hpp"
+#include "traits.hpp"
+#include "common.hpp"
+#include "fwd.hpp"
+#include "defines.hpp"
+
+#include <Kokkos_Core_fwd.hpp>
 
 #if defined(KOKKOS_ENABLE_OPENMP)
 #  include <omp.h>
