+++ src/variants/view.cpp.in
@@ -1,6 +1,9 @@
 // clang-format off
 
-#include "libpykokkos.hpp"
+#include "traits.hpp"
+#include "common.hpp"
+#include "fwd.hpp"
+#include "defines.hpp"
 
 @PROTOTYPES@
